pyproject.toml
[project]
name = "opus-target-system"
version = "0.1.0"
description = "Claude Opus Multi-Agent Target System for Red Team Validation"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "anthropic>=0.50.0",
    "chromadb>=0.6.0",
    "sentence-transformers>=3.0",
    "pyyaml>=6.0",
    "httpx>=0.28",
    "python-dotenv>=1.0.0",
    "pydantic>=2.10",
    "faker>=28.0.0",
    "rich>=13.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24", "pytest-httpx>=0.35"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bluestack", "harness"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
